WIP - Refactor detections
